third_party/WebKit/LayoutTests/fast/css/remove-sheet-no-layout-expected.txt

Remove a stylesheet not affecting layout.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(removed).backgroundColor is red
PASS recalcCount is 1
PASS layoutCount is 0
PASS getComputedStyle(removed).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

This text should have a green background.
